# sdk/openapi_patched.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manushya.ai",
    "description": "Secure identity + memory infrastructure for autonomous AI agents",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/monitoring/rate-limits": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Get Rate Limit Status",
        "description": "Get current rate limit status for the tenant.",
        "operationId": "get_rate_limit_status_v1_monitoring_rate_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/monitoring/usage": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Get Usage Metrics",
        "description": "Get comprehensive usage metrics.",
        "operationId": "get_usage_metrics_v1_monitoring_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageMetricsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/monitoring/performance": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Get Performance Metrics",
        "description": "Get performance metrics.",
        "operationId": "get_performance_metrics_v1_monitoring_performance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceMetricsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/monitoring/audit-trail": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Get Audit Trail",
        "description": "Get audit trail with filtering.",
        "operationId": "get_audit_trail_v1_monitoring_audit_trail_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            },
            "description": "Filter by event type"
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor ID",
              "title": "Actor Id"
            },
            "description": "Filter by actor ID"
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resource type",
              "title": "Resource Type"
            },
            "description": "Filter by resource type"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of days to look back",
              "default": 7,
              "title": "Days"
            },
            "description": "Number of days to look back"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Maximum number of records",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of records"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditTrailResponse"
                  },
                  "title": "Response Get Audit Trail V1 Monitoring Audit Trail Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/monitoring/webhook-stats": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Get Webhook Delivery Stats",
        "description": "Get webhook delivery statistics.",
        "operationId": "get_webhook_delivery_stats_v1_monitoring_webhook_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of days to analyze",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of days to analyze"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditTrailResponse": {
        "properties": {
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "before_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "Before State"
          },
          "after_state": {
            "additionalProperties": true,
            "type": "object",
            "title": "After State"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "actor_id",
          "resource_type",
          "resource_id",
          "tenant_id",
          "timestamp",
          "before_state",
          "after_state",
          "ip_address",
          "user_agent"
        ],
        "title": "AuditTrailResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PerformanceMetricsResponse": {
        "properties": {
          "avg_response_time_ms": {
            "type": "number",
            "title": "Avg Response Time Ms"
          },
          "p95_response_time_ms": {
            "type": "number",
            "title": "P95 Response Time Ms"
          },
          "p99_response_time_ms": {
            "type": "number",
            "title": "P99 Response Time Ms"
          },
          "requests_per_second": {
            "type": "number",
            "title": "Requests Per Second"
          },
          "error_rate_percent": {
            "type": "number",
            "title": "Error Rate Percent"
          },
          "database_connections": {
            "type": "integer",
            "title": "Database Connections"
          },
          "redis_memory_usage_mb": {
            "type": "number",
            "title": "Redis Memory Usage Mb"
          },
          "background_task_queue_size": {
            "type": "integer",
            "title": "Background Task Queue Size"
          }
        },
        "type": "object",
        "required": [
          "avg_response_time_ms",
          "p95_response_time_ms",
          "p99_response_time_ms",
          "requests_per_second",
          "error_rate_percent",
          "database_connections",
          "redis_memory_usage_mb",
          "background_task_queue_size"
        ],
        "title": "PerformanceMetricsResponse"
      },
...
      "UsageMetricsResponse": {
        "properties": {
          "total_identities": {
            "type": "integer",
            "title": "Total Identities"
          },
          "active_sessions": {
            "type": "integer",
            "title": "Active Sessions"
          },
          "total_memories": {
            "type": "integer",
            "title": "Total Memories"
          },
          "total_api_keys": {
            "type": "integer",
            "title": "Total Api Keys"
          },
          "total_webhooks": {
            "type": "integer",
            "title": "Total Webhooks"
          },
          "total_invitations": {
            "type": "integer",
            "title": "Total Invitations"
          },
          "rate_limit_violations": {
            "type": "integer",
            "title": "Rate Limit Violations"
          },
          "webhook_delivery_success_rate": {
            "type": "number",
            "title": "Webhook Delivery Success Rate"
          },
          "memory_search_queries": {
            "type": "integer",
            "title": "Memory Search Queries"
          },
          "last_24h_activity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Last 24H Activity"
          }
        },
        "type": "object",
        "required": [
          "total_identities",
          "active_sessions",
          "total_memories",
          "total_api_keys",
          "total_webhooks",
          "total_invitations",
          "rate_limit_violations",
          "webhook_delivery_success_rate",
          "memory_search_queries",
          "last_24h_activity"
        ],
        "title": "UsageMetricsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ]
}